+++ docs/changes.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Version 0.5.0
+'''''''''''''
+
+To be released.
+
+
 Version 0.4.0
 '''''''''''''